Made bunch of changes
 piercing shot should work now
lose screen is in
fixed that one issue with the ole shaky skeleboys (it was the bullet, i forgot to turn it off in the editor).
The ice thing is the just old construct instituting it's own CD, so I'll update the tooltip to say that there's a CD on it
I don't see an issue with the bob encounter and the inability to shoot, what abilities were you using?
Multishot wasn't playing the shoot sound when you just had multishot (or dash as well), but that's fixed.
Added a dark background to the story texts
increased text size of the card tooltips
Fixed the boss music issue, the boss music will end when you get back to the lobby (so it'll still play boss music during the story stuff afterwards.